panthenot.md: Panthenot is dead. Long live Panthenot!
The problem (for me) with Pantheon, are the settings-restricted apps,
and the backwards, intrusive overrides (affecting other installed DE's).
It is not a difficult battle(dconf or gschema edits), but an annoying one to fight.

Fedora 30 includes an un-override version of Pantheon that is fully
functional, and far more cohesive than the gala-over-xfce Panthenot attempt.
Fedora's version also allows a fairly easy installation of the session
without elementary's apps.
All of the win, none of the loss (except that it is Fed, not oS).